// src/types/bankBranch.ts
type strNum = string | number;

export interface IBankBranch {
  GPS_X: strNum;
  GPS_Y: strNum;
  bel_number_schet: string;
  cbu_num: string;
  dop_num: string;
  filial_id: strNum;
  filial_name: string;
  filial_num: strNum;
  foreign_number_schet: string;
  home_number: string;
  home_number_prev: string;
  info_bank_bik: string;
  info_bank_unp: string;
  info_text: string;
  info_weekend1_day: string;
  info_weekend1_time: string;
  info_weekend2_day: string;
  info_weekend2_time: string;
  info_weekend3_day: string;
  info_weekend3_time: string;
  info_weekend4_day: string;
  info_weekend4_time: string;
  info_weekend5_day: string;
  info_weekend5_time: string;
  info_weekend6_day: string;
  info_weekend6_time: string;
  info_weekend7_day: string;
  info_weekend7_time: string;
  info_worktime: string;
  name: string;
  name_prev: string | null;
  name_type: string;
  name_type_prev: string | null;
  otd_num: string;
  phone_info: string;
  sap_id: strNum;
  street: string;
  street_prev: string;
  street_type: string;
  street_type_prev: string;
  usl_broker: strNum;
  usl_buy_slitki: strNum;
  usl_card_internet: strNum;
  usl_cennie_bumagi: strNum;
  usl_check_dover_vnebanka: strNum;
  usl_cheki_gilie: strNum;
  usl_cheki_imuschestvo: strNum;
  usl_club_barhat: strNum;
  usl_club_kartblansh: strNum;
  usl_club_ledi: strNum;
  usl_club_nastart: strNum;
  usl_club_persona: strNum;
  usl_club_schodry: strNum;
  usl_club_svoi: strNum;
  usl_club_zclass: strNum;
  usl_coins_exchange: strNum;
  usl_dep_doverennosti: strNum;
  usl_dep_scheta: strNum;
  usl_dep_viplati: strNum;
  usl_depositariy: strNum;
  usl_docObligac_belarusbank: strNum;
  usl_dover_upr: strNum;
  usl_dover_upr_gos: strNum;
  usl_drag_metal: strNum;
  usl_ibank: strNum;
  usl_inkasso_priem: strNum;
  usl_inkasso_priem_deneg_bel: strNum;
  usl_int_cards: strNum;
  usl_izbiz_scheta_operacii: strNum;
  usl_izbiz_scheta_otkr: strNum;
  usl_kamni_brill: strNum;
  usl_konversiya_foreign_val: strNum;
  usl_loterei: strNum;
  usl_mo_rb: strNum;
  usl_operations_bezdokumentar_obligacii: strNum;
  usl_operations_sber_sertif: strNum;
  usl_perechislenie_po_rekvizitam_kartochki: strNum;
  usl_perechislenie_so_scheta_bez_kart: strNum;
  usl_plategi: strNum;
  usl_plategi_all: strNum;
  usl_plategi_in_foreign_val: strNum;
  usl_plategi_minus_internet: strNum;
  usl_plategi_minus_mobi: strNum;
  usl_plategi_minus_mobi_internet_full: strNum;
  usl_plategi_nal_minus_krome_kredit: strNum;
  usl_plategi_za_proezd_v_polzu_banka: strNum;
  usl_podlinnost_banknot: strNum;
  usl_pogashenie_documentar_obligacii: strNum;
  usl_popolnenieSchetaBezKart: strNum;
  usl_popolnenieSchetaBynIspKarts: strNum;
  usl_popolnenieSchetaUsdIspKarts: strNum;
  usl_pov: strNum;
  usl_priemDocPokupkaObl: strNum;
  usl_priem_cennostei_na_hranenie: strNum;
  usl_priem_cennostej_na_hranenie: strNum;
  usl_priem_doc_na_kredits_overdrafts: strNum;
  usl_priem_doc_na_lizing: strNum;
  usl_priem_docs_fl_depozit_operations: strNum;
  usl_priem_docs_vidacha_sopr_lgot_ipotech: strNum;
  usl_priem_inkasso: strNum;
  usl_priem_obl_mf: strNum;
  usl_priem_platejei_byn_ip: strNum;
  usl_priem_platejei_eur_ip: strNum;
  usl_priem_vznosov_inostr_val_ot_strax_agentov: strNum;
  usl_priem_zayvleniy_obsluzhivanie_derzhatelej: strNum;
  usl_prodaga_monet: strNum;
  usl_razm_prodazha_documentar_obligacii: strNum;
  usl_razmen_foreign_val: strNum;
  usl_rb_card: strNum;
  usl_registration_val_dogovor: strNum;
  usl_return_BynIspKarts: strNum;
  usl_return_UsdIspKarts: strNum;
  usl_rko: strNum;
  usl_seif: strNum;
  usl_soprov_kredit_v_tom_chisle_magnit: strNum;
  usl_strahovanie_avto: strNum;
  usl_strahovanie_avto_pogran: strNum;
  usl_strahovanie_detei: strNum;
  usl_strahovanie_dohod_pod_zaschitoy: strNum;
  usl_strahovanie_express: strNum;
  usl_strahovanie_green_karta: strNum;
  usl_strahovanie_home: strNum;
  usl_strahovanie_kartochki: strNum;
  usl_strahovanie_kasko: strNum;
  usl_strahovanie_komplex: strNum;
  usl_strahovanie_medicine_nerezident: strNum;
  usl_strahovanie_perevozki: strNum;
  usl_strahovanie_s_zabotoi_o_blizkih: strNum;
  usl_strahovanie_timeAbroad: strNum;
  usl_strahovanie_zashhita_ot_kleshha: strNum;
  usl_strahovka_site: strNum;
  usl_stroysber: strNum;
  usl_stroysber_new: strNum;
  usl_subsidiya_scheta: strNum;
  usl_swift: strNum;
  usl_vidach_spravok_po_kredit_overdr: strNum;
  usl_vklad: strNum;
  usl_vozvrat_nds: strNum;
  usl_vydacha_nal_v_banke: strNum;
  usl_vydacha_vypiski: strNum;
  usl_vypllata_bel_rub: strNum;
  usl_vzk: strNum;
}